requirements.txt
Flask==3.0.1
Flask-CORS==4.0.0
tensorflow==2.15.0
Pillow==9.4.0
numpy==1.25.0
matplotlib==3.7.2
